2023/Q4/React_Django_Chat_Application/chat_app_backend/backend/requirements.txt
asgiref==3.7.2
async-timeout==4.0.3
channels==4.0.0
channels-redis==4.1.0
Django==5.0
django-cors-headers==4.3.1
djangorestframework==3.14.0
install==1.3.5
msgpack==1.0.7
PyJWT==2.8.0
pytz==2023.3.post1
redis==5.0.1
sqlparse==0.4.4
typing_extensions==4.8.0
